reorganized matrix test utilities
