+++ test/include/dlaf_test/matrix/util_matrix.h
@@ -48,88 +48,6 @@ void set(MatrixType<T, Device::CPU>& mat, ElementGetter el) {
   }
 }
 
-/// @brief Returns a col-major ordered vector with the futures to the matrix tiles.
-template <template <class, Device> class MatrixType, class T, Device device>
-std::vector<hpx::future<Tile<T, device>>> getFuturesUsingLocalIndex(MatrixType<T, device>& mat) {
-  const matrix::Distribution& dist = mat.distribution();
-
-  std::vector<hpx::future<Tile<T, device>>> result;
-  result.reserve(util::size_t::mul(dist.localNrTiles().rows(), dist.localNrTiles().cols()));
-
-  for (SizeType j = 0; j < dist.localNrTiles().cols(); ++j) {
-    for (SizeType i = 0; i < dist.localNrTiles().rows(); ++i) {
-      result.emplace_back(std::move(mat(LocalTileIndex(i, j))));
-      EXPECT_TRUE(result.back().valid());
-    }
-  }
-  return result;
-}
-
-/// @brief Returns a col-major ordered vector with the futures to the matrix tiles.
-template <template <class, Device> class MatrixType, class T, Device device>
-std::vector<hpx::future<Tile<T, device>>> getFuturesUsingGlobalIndex(MatrixType<T, device>& mat) {
-  const matrix::Distribution& dist = mat.distribution();
-
-  std::vector<hpx::future<Tile<T, device>>> result;
-  result.reserve(util::size_t::mul(dist.localNrTiles().rows(), dist.localNrTiles().cols()));
-
-  for (SizeType j = 0; j < dist.nrTiles().cols(); ++j) {
-    for (SizeType i = 0; i < dist.nrTiles().rows(); ++i) {
-      GlobalTileIndex global_index{i, j};
-      comm::Index2D owner = dist.rankGlobalTile(global_index);
-
-      if (dist.rankIndex() == owner) {
-        result.emplace_back(std::move(mat(global_index)));
-        EXPECT_TRUE(result.back().valid());
-      }
-    }
-  }
-  return result;
-}
-
-/// @brief Returns a col-major ordered vector with the read-only shared-futures to the matrix tiles.
-template <template <class, Device> class MatrixType, class T, Device device>
-std::vector<hpx::shared_future<Tile<const T, device>>> getSharedFuturesUsingLocal(
-    MatrixType<T, device>& mat) {
-  const matrix::Distribution& dist = mat.distribution();
-
-  std::vector<hpx::shared_future<Tile<const T, device>>> result;
-  result.reserve(util::size_t::mul(dist.localNrTiles().rows(), dist.localNrTiles().cols()));
-
-  for (SizeType j = 0; j < dist.localNrTiles().cols(); ++j) {
-    for (SizeType i = 0; i < dist.localNrTiles().rows(); ++i) {
-      result.emplace_back(mat.read(LocalTileIndex(i, j)));
-      EXPECT_TRUE(result.back().valid());
-    }
-  }
-
-  return result;
-}
-
-/// @brief Returns a col-major ordered vector with the read-only shared-futures to the matrix tiles.
-template <template <class, Device> class MatrixType, class T, Device device>
-std::vector<hpx::shared_future<Tile<const T, device>>> getSharedFuturesUsingGlobal(
-    MatrixType<T, device>& mat) {
-  const matrix::Distribution& dist = mat.distribution();
-
-  std::vector<hpx::shared_future<Tile<const T, device>>> result;
-  result.reserve(util::size_t::mul(dist.localNrTiles().rows(), dist.localNrTiles().cols()));
-
-  for (SizeType j = 0; j < dist.nrTiles().cols(); ++j) {
-    for (SizeType i = 0; i < dist.nrTiles().rows(); ++i) {
-      GlobalTileIndex global_index{i, j};
-      comm::Index2D owner = dist.rankGlobalTile(global_index);
-
-      if (dist.rankIndex() == owner) {
-        result.emplace_back(mat.read(global_index));
-        EXPECT_TRUE(result.back().valid());
-      }
-    }
-  }
-
-  return result;
-}
-
 /// @brief Checks the elements of the matrix.
 ///
 /// comp(expected({i, j}), (i, j)-element) is used to compare the elements.
